pyproject.toml
[tool.poetry]
name = "x-trade"
version = "0.1.0"
description = ""
authors = ["saixfix <maxim-grishins@mail.ru>"]
readme = "README.md"
packages = [{include = "x_trade"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
